Comment out HCR plots
Low utility, and they don't seem to be working. For example, Figure 39 and 40 (PRD MPs for DDM and DIM respectively) seem to show the exact same data. In addition, the HR for some panels goes up to 1.0 (see "HS30-60 HR10 cap2.5" in Figure 39) -- this seems to be just for some of the MPs with "HR10".