most png sent to herringsr_figures folder
For this to work a herringsr_figures folder must be created in the same location the herringsr repo is on one's local computer.   This was done in consultation with JC to facilitate a request by MG, so that the ppt figures don't take up repo space. Additions to the gitignore were also made. There are some pngs that are to stay in the herringsr/figures folder (not the herringsr_figures folder)